Add test files for all lib files (#177)
This adds in test files with pending tests that I'm intending on writing
in separate PRs. I've written tests for some of the simpler stuff as
part of this to give an idea of how the newer tests will be written.

I'm making extensive use of mocking, and the intention is to also
produce a suite of integration tests later which will test the whole
module with an example Express application.

This starts to resolve #42.