Allow setting API keys through the options object (#184)
* Allow setting API keys through the options object

This mostly means that apps no longer need to specify an
FT_GRAPHITE_APIKEY environment variable just to turn it off. We do this
in Origami Services currently.

Also it feels wrong that every instance of Metrics has to share the same
API keys.

* Add some unit tests

* Clarify the missing key messages

* Coerce keys from environment variables to false

This is a little neater where we check the keys, but complicates the
setting of them a little. I think it's probably better this way

* Fix a lint error