Remove non-existent services
While reviewing a PR I noticed that we have a lot of services listed
here which either don't have DNS or are Heroku apps that no longer
exists.

We should remove these from this file because we actually want alerts
to go off if we're connecting to these services - it would indicate
a problem.

This is probably a breaking change.