Add api-t.ft.com/graphql/v1/api to services
next-profile is requesting this service which is causing its healthcheck
to fail intermittently.

https://heimdall.ftops.tech/system?code=next-profile

I've done this in a slightly different way because this new structured
format might allow us in future to add tests to ensure that (for example)
the domain has a DNS record or provide some example URLs which we can verify
work properly.

It also means that developers don't have to write regexes frequently anymore.