Make the subs graphql match on different paths
The Subs GraphQL can be queried on either

https://api.ft.com/subs/query/graphql
https://api.ft.com/subs/query/

To allow for this the service regex needs to be more flexible whilst not matching on

https://api.ft.com/subs/query/api

which is used for canned queries.

The regex additions added in this commit are

(\/graphql)? = This makes it optional for the path to end with /graphql
(\/) ? = This makes it optional for it to end with a trailing slaash
$ = This means that this must be the end of the path/line

This means the following will all match

https://api.ft.com/subs/query/graphql
https://api.ft.com/subs/query/graphql/
https://api.ft.com/subs/query/

Whilst

https://api.ft.com/subs/query/api

Does not match and will fall through to the regex aimed at that path

By using the end of line check ($) we are assuming that no extra details will be added to the path
such as query params or hashes. This would not currently be common place when the POST requests
being made to query these endpoints